# openapi.json
{"openapi":"3.1.0","info":{"title":"Meow Mingle API","description":"A cat dating app to find your purrfect match","license":{"name":"MIT","identifier":"MIT"},"version":"0.1.0"},"paths":{"/api/v1/auth/sign-in":{"post":{"tags":["crate::handlers::auth"],"operationId":"sign_in_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignInPayload"}}},"required":true},"responses":{"200":{"description":"Auth sign in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignInResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/auth/sign-up":{"post":{"tags":["crate::handlers::auth"],"operationId":"sign_up_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignUpPayload"}}},"required":true},"responses":{"200":{"description":"Auth sign up","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignUpResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/cats":{"get":{"tags":["crate::handlers::cats"],"operationId":"cats_list_handler","responses":{"200":{"description":"List of cats","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatsListResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/cats/{name}":{"get":{"tags":["crate::handlers::cats"],"operationId":"cat_detail_handler","parameters":[{"name":"name","in":"path","description":"Cat name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Details for a specific cat","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatDetailResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/matches":{"get":{"tags":["crate::handlers::matches"],"operationId":"matches_list_handler","responses":{"200":{"description":"List of all matches for a specific cat","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchesListResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/matches/suggestions":{"get":{"tags":["crate::handlers::matches"],"operationId":"match_suggestions_handler","responses":{"200":{"description":"List of match suggestions for a specific cat","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchSuggestionsResponse"}}}},"500":{"description":"Internal server error"}}}},"/api/v1/session/{id}":{"get":{"tags":["crate::handlers::session"],"operationId":"session_get_by_id_handler","parameters":[{"name":"id","in":"path","description":"Session id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Details for a specific cat","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatDetailResponse"}}}},"500":{"description":"Internal server error"}}}}},"components":{"schemas":{"AuthSignInPayload":{"type":"object","required":["name","password"],"properties":{"name":{"type":"string"},"password":{"type":"string"}}},"AuthSignInResponse":{"type":"object","required":["status","message"],"properties":{"message":{"type":"string"},"status":{"$ref":"#/components/schemas/Status"}}},"AuthSignUpPayload":{"type":"object","required":["cat"],"properties":{"cat":{"$ref":"#/components/schemas/NewCat"}}},"AuthSignUpResponse":{"type":"object","required":["status","message"],"properties":{"message":{"type":"string"},"results":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/AuthSignUpResponseResults"}]},"status":{"$ref":"#/components/schemas/Status"}}},"AuthSignUpResponseResults":{"type":"object","required":["cat","session_id"],"properties":{"cat":{"$ref":"#/components/schemas/Cat"},"session_id":{"type":"string"}}},"Cat":{"type":"object","required":["id","name"],"properties":{"active":{"type":["boolean","null"]},"age":{"type":["integer","null"],"format":"int32"},"avatar_filename":{"type":["string","null"]},"biography":{"type":["string","null"]},"breed_id":{"type":["string","null"],"format":"uuid"},"breed_name":{"type":["string","null"]},"created_at":{"type":["string","null"],"format":"date-time"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"updated_at":{"type":["string","null"],"format":"date-time"}}},"CatDetailResponse":{"type":"object","required":["status"],"properties":{"message":{"type":["string","null"]},"results":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Cat"}]},"status":{"$ref":"#/components/schemas/Status"}}},"CatsListResponse":{"type":"object","required":["status","results"],"properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/Cat"}},"status":{"type":"string"}}},"Match":{"type":"object","required":["id","initiator_id","target_id","status"],"properties":{"id":{"type":"string","format":"uuid"},"initiator_id":{"type":"string","format":"uuid"},"status":{"$ref":"#/components/schemas/MatchStatus"},"target_id":{"type":"string","format":"uuid"}}},"MatchStatus":{"type":"string","enum":["Pending","Accepted","Declined"]},"MatchSuggestionsResponse":{"type":"object","required":["status","results"],"properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/Cat"}},"status":{"type":"string"}}},"MatchesListResponse":{"type":"object","required":["status","results"],"properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/Match"}},"status":{"type":"string"}}},"NewCat":{"type":"object","required":["name","password","breed_id"],"properties":{"age":{"type":["integer","null"],"format":"int32"},"breed_id":{"type":"string","format":"uuid"},"name":{"type":"string"},"password":{"type":"string"}}},"Status":{"type":"string","enum":["OK","ERROR"]}}},"tags":[{"name":"cats","description":"Cat endpoints"},{"name":"matches","description":"Match endpoints"},{"name":"auth","description":"Auth endpoints"}]}